+++ cedar-policy/src/ffi/is_authorized.rs
@@ -881,7 +881,7 @@ mod test {
              }
             },
             "policies": {
-                "staticPolicies": "permit(principal == User::\"alice\", action, resource) when { context.is_authenticated && context.source_ip.isInRange(ip(\"222.222.222.0/24\")) };"
+                "staticPolicies": "permit(principal == User::\"alice\", action, resource) when { context.is_authenticated && context.source_ip.isInRange(ip(\"192.167.0.1/24\"), ip(\"222.222.222.0/24\")) };"
             },
             "entities": []
         });